+++ test/libsolidity/syntaxTests/nameAndTypeResolution/043_functions_with_stucts_of_non_external_types_in_interface_nested.sol
@@ -7,4 +7,5 @@ contract C {
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
+// TypeError: (132-140): Mappings cannot live outside storage.
 // TypeError: (132-140): Internal or recursive type is not allowed for public or external functions.