+++ test/libsolidity/syntaxTests/types/mapping/library_array_argument_external.sol
@@ -3,4 +3,4 @@ library L {
     }
 }
 // ----
-// TypeError: (27-58): Type is required to live outside storage.
+// TypeError: (27-58): Mappings cannot live outside storage.